moved files to respective folders and relinked the main.js and index.html page
