+++ index.html
@@ -8,6 +8,12 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/main.js"></script>
+    <script type="module" src="/js/counter.js"></script>
+    <script type="module" src="/js/main.js"></script>
+    <script type="module" src="/images/javascript.svg"></script>
+    <script type="module" src="/images/vite.svg"></script>
+    <script type="module" src="/css/style.css"></script>
+    <script type="module" src="/package.json"></script>
+    <script type="module" src="/package-lock.json"></script>
   </body>
 </html>
