+++ js/main.js
@@ -1,7 +1,7 @@
-import './style.css'
-import javascriptLogo from './javascript.svg'
-import viteLogo from '/vite.svg'
-import { setupCounter } from './counter.js'
+import '/css/style.css'
+import javascriptLogo from '/images/javascript.svg'
+import viteLogo from '/images/vite.svg'
+import { setupCounter } from '/js/counter.js'
 import dayjs from 'dayjs'
 
 const startDate = dayjs('2023-09-05');
