+++ package.json
@@ -6,6 +6,7 @@
   "license": "MIT",
   "author": "Sterling Smith <sterling12311@gmail.com>",
   "type": "module",
+  "main": "index.html",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
